+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+# [v1.7.0] (Unreleased)
+### Notes
+- This release supports API2400 minimally where we can use OneView v5.60 with this SDK. 
+
+### Oneview Features supported
+- Hypervisor Manager
+- Hypervisor Cluster Profile
+- Scope
+- Server Certificate
+- Storage Pool
+- Storage Volume
+- Storage Volume Attachments
+- Storage Volume Template
+- Storage System
+- Uplink Set
+
 # [v1.6.0]
 ### Notes
 - This release supports API2200 minimally where we can use OneView API2200 with this SDK and ImageStreamer REST API version 2000 (I3S v5.40).